+++ public/pages/DetectorDetail/containers/DetectorDetail.tsx
@@ -1,4 +1,3 @@
-
 /*
  * SPDX-License-Identifier: Apache-2.0
  *
