+++ public/redux/reducers/ad.ts
@@ -18,7 +18,7 @@ import {
 import handleActions from '../utils/handleActions';
 import { Detector, DetectorListItem } from '../../models/interfaces';
 import { AD_NODE_API } from '../../../utils/constants';
-import { GetDetectorsQueryParams, MDSQueryParams } from '../../../server/models/types';
+import { GetDetectorsQueryParams } from '../../../server/models/types';
 import { cloneDeep, get } from 'lodash';
 import moment from 'moment';
 import { DETECTOR_STATE } from '../../../server/utils/constants';
